feat: Implement basic cli interface. Parsing for a specific PDF working
This commit add some basic cli interface to parse pdf in the filesystem.

Parsing for a specific PDF is working. Any other PDF isn't.

Concentrate the stream parsing for after the whole PDF is parsed. This
simplify the code complex.
Add some logic to handle the operators that aren't necessary to extract
the PDF text.